// src/models/sharedCartStatus.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * Shop API
 * OpenAPI spec version: v1
 */

export type SharedCartStatus =
  typeof SharedCartStatus[keyof typeof SharedCartStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SharedCartStatus = {
  Active: "Active",
  CheckoutInProgress: "CheckoutInProgress",
  Completed: "Completed",
} as const;